Adjust Flycheck & Formatter options
Flycheck behaviour is as follows:
By default StandardJS will be used, but if ESLint is installed inside a
project, then use ESLint instead.

Formatter behaviour is as follows:
By default StandardJS will be used, but if PrettierJS is installed
inside a project, then use PrettierJS instead and if ESLint is installed
inside a project, then use ESLint instead.