Fix bug
I accidentally was checking, if the wireguard sytem service was up,
which I always started automatically on startup. And even if the
wireguard interface was down, the service was still up & running.
This is useless, of course. Now I am checking, if the wireguard
interface is existing, which is a way better check. If it exists,
than I know it is up & running. This means in simple words:
No interface => no wireguard!